src/app/admin/admin.service.spec.ts
import { TestBed, inject } from '@angular/core/testing';

import { AuthenticationGuard, AuthenticationService } from '@app/auth';
import { MockAuthenticationService } from '@app/auth/authentication.service.mock';
import { AdminComponent } from './admin.component';
import { Admin } from './admin.service';

describe('Admin', () => {
  beforeEach(() => {
    TestBed.configureTestingModule({
      declarations: [AdminComponent],
      providers: [AuthenticationGuard, { provide: AuthenticationService, useClass: MockAuthenticationService }],
    });
  });

  describe('childRoutes', () => {
    it('should create routes as children of admin', () => {
      // Prepare
      const testRoutes = [{ path: 'test' }];

      // Act
      const result = Admin.childRoutes(testRoutes);

      // Assert
      expect(result.path).toBe('');
      expect(result.children).toBe(testRoutes);
      expect(result.component).toBe(AdminComponent);
    });
  });
});
